% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_anomalies.R
\name{count_anomalies}
\alias{count_anomalies}
\title{Count Anomalies}
\usage{
count_anomalies(data, col_name, min, max)
}
\arguments{
\item{data}{A data frame, tibble, or `dtplyr::lazy_dt` containing the data.}

\item{col_name}{The column within the data to check for anomalies. This should
be an unquoted column name.}

\item{min}{The minimum threshold for determining anomalies.}

\item{max}{The maximum threshold for determining anomalies.}
}
\value{
A tibble containing the number and percentage of short and long rides.
}
\description{
Calculate the number and percentage of anomalies in a specified column based
on given minimum and maximum thresholds.
}
\details{
This function calculates the number and percentage of values in a
  specified column that fall below the minimum threshold (`min`) or exceed the
  maximum threshold (`max`). It returns a tibble with these counts and
  percentages.
}
\examples{
\dontrun{
library(dplyr)
library(tibble)

data <- data.frame(ride_length = c(5, 10, 15, 20, 25, 30, 35, 40, 45, 50))
result <- count_anomalies(data, ride_length, min = 10, max = 40)
print(result)
}

}
