% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_categorical.R
\name{count_categorical}
\alias{count_categorical}
\title{Count Categorical Variables}
\usage{
count_categorical(data, categorical_vars)
}
\arguments{
\item{data}{A data frame, tibble, or `dtplyr::lazy_dt` containing the data.}

\item{categorical_vars}{The categorical variable to group by. This should
be unquoted variable names.}
}
\value{
A tibble containing the counts of each level of the specified
  categorical variable.
}
\description{
Calculate the count of each level for specified categorical variables.
}
\details{
This function groups the data by the specified categorical variable
  and calculates the count of each level. It returns a tibble with these
  counts.
}
\examples{
\dontrun{
library(dplyr)
library(tibble)

data <- data.frame(
  category1 = c("A", "B", "A", "B", "C"),
  category2 = c("X", "X", "Y", "Y", "Z")
)
result <- count_categorical(data, c(category1, category2))
print(result)
}

}
