% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_lazy_dt.R
\name{csv_to_lazy_dt}
\alias{csv_to_lazy_dt}
\title{Import and Combine CSV Files into a Lazy Data Table}
\usage{
csv_to_lazy_dt()
}
\value{
A `dtplyr::lazy_dt` object containing the combined data from all
  CSV files in the specified directory.
}
\description{
Read CSV files from the specified directory and convert them into a
`dtplyr::lazy_dt` object.
}
\details{
This function reads all CSV files from the "data/raw_data" directory,
  combines them into a single data frame, and then converts it into a
  `dtplyr::lazy_dt` object.
}
\examples{
\dontrun{
library(dplyr)
library(dtplyr)
library(vroom)
library(here)

lazy_data <- csv_to_lazy_dt()
print(lazy_data)
}

}
