% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mode.R
\name{find_mode}
\alias{find_mode}
\title{Find Mode}
\usage{
find_mode(data, variable)
}
\arguments{
\item{data}{A data frame, tibble, or `dtplyr::lazy_dt` containing the data.}

\item{variable}{The variable within the data for which to find the mode.
This should be an unquoted variable name.}
}
\value{
The mode(s) of the specified variable.
}
\description{
Identify the mode(s) of a specified variable in a data frame.
}
\details{
This function calculates the mode(s) of a given variable by counting
  the frequency of each value and identifying the value(s) with the highest
  frequency.
}
\examples{
\dontrun{
library(dplyr)

data <- data.frame(
  values = c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4)
)
mode_value <- find_mode(data, values)
print(mode_value)
}

}
