% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow_lazy_dt.R
\name{nrow_lazy_dt}
\alias{nrow_lazy_dt}
\title{Count the Number of Observations in a Lazy Data Table}
\usage{
nrow_lazy_dt(data)
}
\arguments{
\item{data}{A `dtplyr::lazy_dt` object containing the data.}
}
\value{
An integer representing the number of observations in the lazy data table.
}
\description{
This function counts the number of observations in a lazy data table,
specifically designed to handle data tables managed by `dtplyr` in a lazy
evaluation mode. This is particularly useful for large datasets that are not
entirely loaded into memory.
}
\details{
The function uses `dplyr`'s `summarize()` to compute the count of
         rows, which is then explicitly collected and extracted using
         `collect()` and `pull()`. This method ensures that the counting
         operation is deferred until explicitly demanded, which optimizes
         memory usage and computational efficiency for large datasets.
}
\examples{
\dontrun{
library(dplyr)
library(dtplyr)

lazy_data <- lazy_dt(mtcars)
row_count <- nrow_lazy_dt(lazy_data)
print(row_count)
}

}
