% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_anomalies.R
\name{remove_anomalies}
\alias{remove_anomalies}
\title{Remove Anomalies}
\usage{
remove_anomalies(data, col_name, min, max)
}
\arguments{
\item{data}{A data frame, tibble, or `dtplyr::lazy_dt` containing the data.}

\item{col_name}{The column within the data to filter for anomalies. This
should be an unquoted column name.}

\item{min}{The minimum threshold for filtering anomalies.}

\item{max}{The maximum threshold for filtering anomalies.}
}
\value{
A filtered data frame, tibble, or `dtplyr::lazy_dt` with anomalies removed.
}
\description{
Filter out anomalies in a specified column based on given minimum and maximum
thresholds.
}
\details{
This function filters out rows where the specified column's values
  are below the minimum threshold (`min`) or above the maximum threshold (`max`).
}
\examples{
\dontrun{
library(dplyr)

data <- data.frame(
  ride_length = c(5, 10, 15, 20, 25, 30, 35, 40, 45, 50)
)
cleaned_data <- remove_anomalies(data, ride_length, min = 10, max = 40)
print(cleaned_data)
}

}
