% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plots.R
\name{save_plots}
\alias{save_plots}
\title{Save Plots}
\usage{
save_plots(filename, plot)
}
\arguments{
\item{filename}{A character string representing the name of the file (without
extension) to save the plot as.}

\item{plot}{A ggplot object to be saved.}
}
\value{
None. This function saves the plot to a file and does not return a value.
}
\description{
Save a ggplot object as a PNG file with specified dimensions and resolution.
}
\details{
This function saves a ggplot object as a PNG file in the "output/plots"
  directory with a specified width, height, and resolution. The file name is
  constructed by appending ".png" to the provided filename.
}
\examples{
\dontrun{
library(ggplot2)
library(here)

plot <- ggplot(mtcars, aes(x = mpg, y = hp)) +
  geom_point()
save_plots("mtcars_plot", plot)
}

}
