% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_tables.R
\name{style_tables}
\alias{style_tables}
\title{Style Tables}
\usage{
style_tables(kable_input, tbl_format)
}
\arguments{
\item{kable_input}{A kable object to be styled.}

\item{tbl_format}{A character string specifying the format of the table.
Accepted values are "html" and "latex".}
}
\value{
A styled kable object.
}
\description{
Apply styling to kable tables based on the specified format.
}
\details{
This function applies different styles to kable tables based on the
  specified format. For HTML tables, it adds bootstrap options such as "hover",
  "condensed", and "responsive". For LaTeX tables, it adds options like
  "striped" and "hold_position".
}
\examples{
\dontrun{
library(knitr)
library(kableExtra)

# Example for HTML table
kable_input <- knitr::kable(head(mtcars), format = "html")
styled_table <- style_tables(kable_input, "html")

# Example for LaTeX table
kable_input <- knitr::kable(head(mtcars), format = "latex")
styled_table <- style_tables(kable_input, "latex")
}

}
